phoenix-api/pyproject.toml
[tool.poetry]
name = "phoenix-api"
version = "0.1.0"
description = "Phoenix API Gateway - Orchestrates product logic and connects to Luna Hub."
authors = ["Dédale <dedale@phoenix-ia.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
httpx = "^0.27.0"
python-dotenv = "^1.0.1"
gunicorn = "^22.0.0"
pydantic = {extras = ["email"], version = "^2.8.2"}
pydantic-settings = "^2.3.4"
google-generativeai = "^0.5.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
ruff = "^0.5.5"
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"